fix(windows): resolve WasapiCapture namespace and pre-empt next-stage compile errors
WasapiCapture lives in NAudio.CoreAudioApi, not NAudio.Wave. Also fixes a
char-to-ushort conversion in the Unicode paste path, an ambiguous
UnhandledExceptionEventArgs between System and Microsoft.UI.Xaml, a
WavFileSource that never cleared IsCapturing so callers polling for the
end of playback would spin, and a top-level-program name clash on 'args'.

+++ windows/src/DoubaoMurmur.App/Platform/PasteHelper.cs
@@ -182,8 +182,8 @@ internal static class PasteHelper
         var inputs = new List<NativeMethods.INPUT>(text.Length * 2);
         foreach (var ch in text)
         {
-            inputs.Add(KeyInput(0, ch, NativeMethods.KEYEVENTF_UNICODE));
-            inputs.Add(KeyInput(0, ch, NativeMethods.KEYEVENTF_UNICODE | NativeMethods.KEYEVENTF_KEYUP));
+            inputs.Add(KeyInput(0, (ushort)ch, NativeMethods.KEYEVENTF_UNICODE));
+            inputs.Add(KeyInput(0, (ushort)ch, NativeMethods.KEYEVENTF_UNICODE | NativeMethods.KEYEVENTF_KEYUP));
 
             // SendInput takes an array; chunk it so very long transcriptions do not
             // allocate one huge block and so the target app can keep up.
